+++ argonaut/src/main/scala/io/fintrospect/formats/Argonaut.scala
@@ -62,7 +62,7 @@ object Argonaut extends JsonLibrary[Json, Json] {
   /**
     * Convenience method for creating BodySpecs that just use straight JSON encoding/decoding logic
     */
-  def bodySpec[R](description: Option[String] = None)(implicit encodec: EncodeJson[R], decodec: DecodeJson[R]) =
+  def bodySpec[R](description: String = null)(implicit encodec: EncodeJson[R], decodec: DecodeJson[R]) =
     BodySpec.json(description, this).map(j => JsonFormat.decode[R](j)(decodec), (u: R) => JsonFormat.encode(u)(encodec))
 
   /**