+++ core/src/test/scala/io/fintrospect/util/MultiBodyTypeTest.scala
@@ -12,7 +12,7 @@ import org.scalatest.{FunSpec, Matchers}
 
 class MultiBodyTypeTest extends FunSpec with Matchers {
 
-  val xmlBody = Body.xml(None)
+  val xmlBody = Body.xml()
   val xmlAccepting = MultiBodyType(xmlBody -> Service.mk { req: Request => Future(Response(Ok)) })
 
   describe("MultiBodyType") {